refactor(clawdstrike-policy-event): extract kinds/policy_decision from edr/receipt/mod.rs
