+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/kinds/mod.rs
@@ -1,11 +1,13 @@
 mod detection;
 mod evidence_bundle;
 mod graph_slice;
+mod policy_decision;
 mod sensor_state;
 mod simulation;
 
 pub(crate) use detection::*;
 pub(crate) use evidence_bundle::*;
 pub(crate) use graph_slice::*;
+pub(crate) use policy_decision::*;
 pub(crate) use sensor_state::*;
 pub(crate) use simulation::*;
